+++ src/main/java/pl/kniotes/elearn_course_viewer/Controllers/PathController.java
@@ -1,17 +1,23 @@
 package pl.kniotes.elearn_course_viewer.Controllers;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import pl.kniotes.elearn_course_viewer.Models.Course.Course;
+import pl.kniotes.elearn_course_viewer.Models.Course.CourseParserJSON;
 import pl.kniotes.elearn_course_viewer.Models.Path.OrderedCourse;
 import pl.kniotes.elearn_course_viewer.Models.Path.Path;
 import pl.kniotes.elearn_course_viewer.Models.Path.Section;
+import pl.kniotes.elearn_course_viewer.Models.Path.SectionParserJSON;
 import pl.kniotes.elearn_course_viewer.Repositories.CourseRepository;
 import pl.kniotes.elearn_course_viewer.Repositories.PathRepository;
 
+import java.lang.reflect.Type;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -40,7 +46,9 @@ public class PathController {
         } else {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Name is not set");
         }
-        path.setDescription(description);
+        if (description != null && !description.isBlank()) {
+            path.setDescription(description);
+        }
         return this.pathRepository.save(path);
     }
 
@@ -67,13 +75,16 @@ public class PathController {
         if (description != null && !description.isBlank()) {
             path.setDescription(description);
         }
-
         return this.pathRepository.save(path);
     }
 
     @CrossOrigin("*")
     @DeleteMapping("/api/paths/{id}")
     public Boolean deletePath(@PathVariable("id") String id) {
+        Path path = this.pathRepository.findPathById(id);
+        if (path == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
+        }
         this.pathRepository.deletePathById(id);
         return this.courseRepository.findCourseById(id) == null;
     }
@@ -90,7 +101,7 @@ public class PathController {
 
     @CrossOrigin("*")
     @PostMapping("/api/paths/{pathId}/sections")
-    public Section postSection(@PathVariable("pathId") String pathId, String name, String description) {
+    public Section postSection(@PathVariable("pathId") String pathId, String name, String description, Integer orderKey) {
         Path path = this.pathRepository.findPathById(pathId);
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
@@ -104,7 +115,26 @@ public class PathController {
         if (description != null && !description.isBlank()) {
             section.setDescription(description);
         }
-        Integer key = path.getSections().size();
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
+        int key = 0;
+        if (orderKey != null) {
+            if (orderKey < 1) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "OrderKey must be bigger than 1");
+            }
+            if (path.getSections().containsKey(orderKey)) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There is a section assigned to this orderKey");
+            }
+            key = orderKey;
+        } else {
+            for (int item : path.getSections().keySet()) {
+                if (key < item) {
+                    key = item;
+                }
+            }
+            key++;
+        }
         path.getSections().put(key, section);
         this.pathRepository.save(path);
         return this.pathRepository.findPathById(pathId).getSections().get(key);
@@ -117,6 +147,9 @@ public class PathController {
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
         }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
         return path.getSections().get(sectionKey);
     }
 
@@ -126,6 +159,9 @@ public class PathController {
         Path path = this.pathRepository.findPathById(pathId);
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
+        }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
         }
         Section section = path.getSections().get(sectionKey);
         if (section == null) {
@@ -138,7 +174,9 @@ public class PathController {
                 throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section name seems to be empty. Set the name in order to update");
             }
         }
-        section.setDescription(description);
+        if (description != null && !description.isBlank()) {
+            section.setDescription(description);
+        }
         this.pathRepository.save(path);
         return this.pathRepository.findPathById(pathId).getSections().get(sectionKey);
     }
@@ -150,6 +188,9 @@ public class PathController {
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
         }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
         if (path.getSections().get(sectionKey) == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section is invalid");
         }
@@ -160,10 +201,23 @@ public class PathController {
 
     @CrossOrigin("*")
     @PutMapping("/api/paths/{pathId}/sections")
-    public Path changeSectionsOrder(@PathVariable("pathId") String pathId, HashMap<Integer, Section> sections) {
+    public Path changeSectionsOrder(@PathVariable("pathId") String pathId, @RequestBody String json) {
         Path path = this.pathRepository.findPathById(pathId);
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
+        }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
+        HashMap<Integer, Section> sections = new HashMap<>();
+        Type placeholderType = new TypeToken<ArrayList<SectionParserJSON>>() {
+        }.getType();
+        List<SectionParserJSON> objects = new Gson().fromJson(json, placeholderType);
+        for (SectionParserJSON item : objects) {
+            if (path.getSections().get(item.getOldIndex()) == null) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "One of the sections seems to not exist");
+            }
+            sections.put(item.getNewIndex(), path.getSections().get(item.getOldIndex()));
         }
         path.setSections(sections);
         return this.pathRepository.save(path);
@@ -175,6 +229,9 @@ public class PathController {
         Path path = this.pathRepository.findPathById(pathId);
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
+        }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
         }
         Section section = path.getSections().get(sectionKey);
         if (section == null) {
@@ -190,15 +247,29 @@ public class PathController {
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
         }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
         Section section = path.getSections().get(sectionKey);
         if (section == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section is invalid");
         }
+        if (section.getCourses() == null) {
+            section.setCourses(new ArrayList<>());
+        }
         int key;
-        if (orderKey != null && !orderKey.toString().isBlank()) {
+        if (orderKey != null) {
+            if (orderKey < 1) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "OrderKey must be bigger than 0");
+            }
+            for (OrderedCourse item : section.getCourses()) {
+                if (item.getOrderIndex() == orderKey) {
+                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There is a course assigned to this orderKey");
+                }
+            }
             key = orderKey;
         } else {
-            key = section.getCourses().size();
+            key = section.getCourses().size() + 1;
         }
         Course course = this.courseRepository.findCourseById(courseId);
         if (course == null) {
@@ -206,23 +277,16 @@ public class PathController {
         }
         OrderedCourse orderedCourse = new OrderedCourse(key, course);
         section.getCourses().add(orderedCourse);
-        this.pathRepository.save(path);
-        return this.pathRepository.findPathById(pathId).getSections().get(sectionKey).getCourses().get(key);
+        path = this.pathRepository.save(path);
+        section = path.getSections().get(sectionKey);
+        OrderedCourse addedCourse = null;
+        for (OrderedCourse item : section.getCourses()) {
+            if (item.getOrderIndex() == key) {
+                addedCourse = item;
+            }
+        }
+        return addedCourse;
     }
-
-//    @CrossOrigin("*")
-//    @GetMapping("/api/paths/{pathId}/sections/{sectionKey}/courses/{courseKey}")
-//    public OrderedCourse getCourseInSection(@PathVariable("pathId") String pathId, @PathVariable("sectionKey") Integer sectionKey, @PathVariable("courseKey") Integer courseKey) {
-//        Path path = this.pathRepository.findPathById(pathId);
-//        if (path == null) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
-//        }
-//        Section section = path.getSections().get(sectionKey);
-//        if (section == null) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section is invalid");
-//        }
-//        return section.getCourses().get(courseKey);
-//    }
 
     @CrossOrigin("*")
     @DeleteMapping("/api/paths/{pathId}/sections/{sectionKey}/courses/{courseKey}")
@@ -231,13 +295,19 @@ public class PathController {
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
         }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
         Section section = path.getSections().get(sectionKey);
         if (section == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section is invalid");
         }
+        if (section.getCourses() == null) {
+            section.setCourses(new ArrayList<>());
+        }
         OrderedCourse orderedCourse = null;
-        for(OrderedCourse item : section.getCourses()) {
-            if(item.getOrderIndex() == courseKey) {
+        for (OrderedCourse item : section.getCourses()) {
+            if (item.getOrderIndex() == courseKey) {
                 orderedCourse = item;
             }
         }
@@ -251,14 +321,30 @@ public class PathController {
 
     @CrossOrigin("*")
     @PutMapping("/api/paths/{pathId}/sections/{sectionKey}/courses")
-    public Section changeCoursesOrderInSection(@PathVariable("pathId") String pathId, @PathVariable("sectionKey") Integer sectionKey, List<OrderedCourse> courses) {
+    public Section changeCoursesOrderInSection(@PathVariable("pathId") String pathId, @PathVariable("sectionKey") Integer sectionKey, @RequestBody String json) {
         Path path = this.pathRepository.findPathById(pathId);
         if (path == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Path is invalid");
         }
+        if (path.getSections() == null) {
+            path.setSections(new HashMap<>());
+        }
         Section section = path.getSections().get(sectionKey);
         if (section == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Section is invalid");
+        }
+        List<OrderedCourse> courses = new ArrayList<>();
+        Type placeholderType = new TypeToken<ArrayList<CourseParserJSON>>() {
+        }.getType();
+        List<CourseParserJSON> objects = new Gson().fromJson(json, placeholderType);
+        for(CourseParserJSON item : objects) {
+            if (this.courseRepository.findCourseById(item.getCourseId()) == null) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "One of the courses seems to not exist");
+            }
+            OrderedCourse course = new OrderedCourse();
+            course.setOrderIndex(item.getOrderIndex());
+            course.setCourse(this.courseRepository.findCourseById(item.getCourseId()));
+            courses.add(course);
         }
         section.setCourses(courses);
         this.pathRepository.save(path);
